go: refactor out `permutations` into own package
This also requires changing 2015/09, since that is where
`permutations` initially was implemented.
