+++ go/internal/permutations/factorial.go
@@ -1,0 +1,9 @@
+package permutations
+
+func factorial(n int) int {
+	if n == 0 {
+		return 1
+	} else {
+		return n * factorial(n-1)
+	}
+}
